// src/types/subsquid/production/v22.ts
export type Type_43 = Type_43_Assets | Type_43_Authorship | Type_43_Babe | Type_43_Balances | Type_43_BridgeMultisig | Type_43_CeresLiquidityLocker | Type_43_CeresStaking | Type_43_Council | Type_43_Currencies | Type_43_DEXAPI | Type_43_Democracy | Type_43_ElectionsPhragmen | Type_43_EthBridge | Type_43_Farming | Type_43_Faucet | Type_43_Grandpa | Type_43_Identity | Type_43_ImOnline | Type_43_IrohaMigration | Type_43_LiquidityProxy | Type_43_MulticollateralBondingCurvePool | Type_43_Multisig | Type_43_Offences | Type_43_Permissions | Type_43_PoolXYK | Type_43_PswapDistribution | Type_43_RandomnessCollectiveFlip | Type_43_Referrals | Type_43_Rewards | Type_43_Scheduler | Type_43_Session | Type_43_Staking | Type_43_Sudo | Type_43_System | Type_43_Technical | Type_43_TechnicalCommittee | Type_43_TechnicalMembership | Type_43_Timestamp | Type_43_TradingPair | Type_43_Utility | Type_43_VestedRewards | Type_43_XSTPool | Type_43_XorFee

export interface Type_43_Assets {
    __kind: 'Assets'
    value: AssetsCall
}

export interface Type_43_Authorship {
    __kind: 'Authorship'
    value: AuthorshipCall
}

export interface Type_43_Babe {
    __kind: 'Babe'
    value: BabeCall
}

export interface Type_43_Balances {
    __kind: 'Balances'
    value: BalancesCall
}

export interface Type_43_BridgeMultisig {
    __kind: 'BridgeMultisig'
    value: BridgeMultisigCall
}

export interface Type_43_CeresLiquidityLocker {
    __kind: 'CeresLiquidityLocker'
    value: CeresLiquidityLockerCall
}

export interface Type_43_CeresStaking {
    __kind: 'CeresStaking'
    value: CeresStakingCall
}

export interface Type_43_Council {
    __kind: 'Council'
    value: CouncilCall
}

export interface Type_43_Currencies {
    __kind: 'Currencies'
    value: CurrenciesCall
}

export interface Type_43_DEXAPI {
    __kind: 'DEXAPI'
    value: DEXAPICall
}

export interface Type_43_Democracy {
    __kind: 'Democracy'
    value: DemocracyCall
}

export interface Type_43_ElectionsPhragmen {
    __kind: 'ElectionsPhragmen'
    value: ElectionsPhragmenCall
}

export interface Type_43_EthBridge {
    __kind: 'EthBridge'
    value: EthBridgeCall
}

export interface Type_43_Farming {
    __kind: 'Farming'
    value: FarmingCall
}

export interface Type_43_Faucet {
    __kind: 'Faucet'
    value: FaucetCall
}

export interface Type_43_Grandpa {
    __kind: 'Grandpa'
    value: GrandpaCall
}

export interface Type_43_Identity {
    __kind: 'Identity'
    value: IdentityCall
}

export interface Type_43_ImOnline {
    __kind: 'ImOnline'
    value: ImOnlineCall
}

export interface Type_43_IrohaMigration {
    __kind: 'IrohaMigration'
    value: IrohaMigrationCall
}

export interface Type_43_LiquidityProxy {
    __kind: 'LiquidityProxy'
    value: LiquidityProxyCall
}

export interface Type_43_MulticollateralBondingCurvePool {
    __kind: 'MulticollateralBondingCurvePool'
    value: MulticollateralBondingCurvePoolCall
}

export interface Type_43_Multisig {
    __kind: 'Multisig'
    value: MultisigCall
}

export interface Type_43_Offences {
    __kind: 'Offences'
    value: OffencesCall
}

export interface Type_43_Permissions {
    __kind: 'Permissions'
    value: PermissionsCall
}

export interface Type_43_PoolXYK {
    __kind: 'PoolXYK'
    value: PoolXYKCall
}

export interface Type_43_PswapDistribution {
    __kind: 'PswapDistribution'
    value: PswapDistributionCall
}

export interface Type_43_RandomnessCollectiveFlip {
    __kind: 'RandomnessCollectiveFlip'
    value: RandomnessCollectiveFlipCall
}

export interface Type_43_Referrals {
    __kind: 'Referrals'
    value: ReferralsCall
}

export interface Type_43_Rewards {
    __kind: 'Rewards'
    value: RewardsCall
}

export interface Type_43_Scheduler {
    __kind: 'Scheduler'
    value: SchedulerCall
}

export interface Type_43_Session {
    __kind: 'Session'
    value: SessionCall
}

export interface Type_43_Staking {
    __kind: 'Staking'
    value: StakingCall
}

export interface Type_43_Sudo {
    __kind: 'Sudo'
    value: SudoCall
}

export interface Type_43_System {
    __kind: 'System'
    value: SystemCall
}

export interface Type_43_Technical {
    __kind: 'Technical'
    value: TechnicalCall
}

export interface Type_43_TechnicalCommittee {
    __kind: 'TechnicalCommittee'
    value: TechnicalCommitteeCall
}

export interface Type_43_TechnicalMembership {
    __kind: 'TechnicalMembership'
    value: TechnicalMembershipCall
}

export interface Type_43_Timestamp {
    __kind: 'Timestamp'
    value: TimestampCall
}

export interface Type_43_TradingPair {
    __kind: 'TradingPair'
    value: TradingPairCall
}

export interface Type_43_Utility {
    __kind: 'Utility'
    value: UtilityCall
}

export interface Type_43_VestedRewards {
    __kind: 'VestedRewards'
    value: VestedRewardsCall
}

export interface Type_43_XSTPool {
    __kind: 'XSTPool'
    value: XSTPoolCall
}

export interface Type_43_XorFee {
    __kind: 'XorFee'
    value: XorFeeCall
}
// ...
export type Weight = bigint

export type LookupSource = Bytes
// ...
export type FaucetCall = FaucetCall_reset_rewards | FaucetCall_transfer
// ...
export type FarmingCall = FarmingCall_migrate_to_1_1
// ...
export type DEXAPICall = DEXAPICall_swap
// ...
export type CurrenciesCall = CurrenciesCall_transfer | CurrenciesCall_transfer_native_currency | CurrenciesCall_update_balance
// ...
export type CouncilCall = CouncilCall_close | CouncilCall_disapprove_proposal | CouncilCall_execute | CouncilCall_propose | CouncilCall_set_members | CouncilCall_vote
// ...
export type CeresStakingCall = CeresStakingCall_deposit | CeresStakingCall_withdraw

export interface CeresStakingCall_deposit {
    __kind: 'deposit'
    amount: Balance
}

export interface CeresStakingCall_withdraw {
    __kind: 'withdraw'
}

export type CeresLiquidityLockerCall = CeresLiquidityLockerCall_change_ceres_fee | CeresLiquidityLockerCall_lock_liquidity

/**
 *  Change CERES fee
 */
export interface CeresLiquidityLockerCall_change_ceres_fee {
    __kind: 'change_ceres_fee'
    ceresFee: Balance
}

/**
 *  Lock liquidity
 */
export interface CeresLiquidityLockerCall_lock_liquidity {
    __kind: 'lock_liquidity'
    assetA: AssetIdOf
    assetB: AssetIdOf
    unlockingBlock: BlockNumber
    percentageOfPoolTokens: Balance
    option: boolean
}

export type BridgeMultisigCall = BridgeMultisigCall_add_signatory | BridgeMultisigCall_approve_as_multi | BridgeMultisigCall_as_multi | BridgeMultisigCall_as_multi_threshold_1 | BridgeMultisigCall_cancel_as_multi | BridgeMultisigCall_register_multisig | BridgeMultisigCall_remove_signatory
// ...
export type BalancesCall = BalancesCall_force_transfer | BalancesCall_set_balance | BalancesCall_transfer | BalancesCall_transfer_keep_alive
// ...
export type BabeCall = BabeCall_report_equivocation | BabeCall_report_equivocation_unsigned
// ...
export type AuthorshipCall = AuthorshipCall_set_uncles

/**
 *  Provide a set of uncles.
 */
export interface AuthorshipCall_set_uncles {
    __kind: 'set_uncles'
    newUncles: Header[]
}

export type AssetsCall = AssetsCall_burn | AssetsCall_mint | AssetsCall_register | AssetsCall_set_non_mintable | AssetsCall_transfer
// ...
export const LookupSource = sts.bytes()

export const Weight = sts.bigint()
// ...
export const CeresStakingCall: sts.Type<CeresStakingCall> = sts.closedEnum(() => {
    return  {
        deposit: sts.enumStruct({
            amount: Balance,
        }),
        withdraw: sts.unit(),
    }
})

export const CeresLiquidityLockerCall: sts.Type<CeresLiquidityLockerCall> = sts.closedEnum(() => {
    return  {
        change_ceres_fee: sts.enumStruct({
            ceresFee: Balance,
        }),
        lock_liquidity: sts.enumStruct({
            assetA: AssetIdOf,
            assetB: AssetIdOf,
            unlockingBlock: BlockNumber,
            percentageOfPoolTokens: Balance,
            option: sts.boolean(),
        }),
    }
})

export const BridgeMultisigCall: sts.Type<BridgeMultisigCall> = sts.closedEnum(() => {
    return  {
        add_signatory: sts.enumStruct({
            newMember: AccountId,
        }),
        approve_as_multi: sts.enumStruct({
            id: AccountId,
            maybeTimepoint: sts.option(() => BridgeTimepoint),
            callHash: sts.bytes(),
            maxWeight: Weight,
        }),
        as_multi: sts.enumStruct({
            id: AccountId,
            maybeTimepoint: sts.option(() => BridgeTimepoint),
            call: OpaqueCall,
            storeCall: sts.boolean(),
            maxWeight: Weight,
        }),
        as_multi_threshold_1: sts.enumStruct({
            id: AccountId,
            call: Type_43,
            timepoint: BridgeTimepoint,
        }),
        cancel_as_multi: sts.enumStruct({
            id: AccountId,
            timepoint: BridgeTimepoint,
            callHash: sts.bytes(),
        }),
        register_multisig: sts.enumStruct({
            signatories: sts.array(() => AccountId),
        }),
        remove_signatory: sts.enumStruct({
            signatory: AccountId,
        }),
    }
})

export const BalancesCall: sts.Type<BalancesCall> = sts.closedEnum(() => {
    return  {
        force_transfer: sts.enumStruct({
            source: LookupSource,
            dest: LookupSource,
            value: sts.bigint(),
        }),
        set_balance: sts.enumStruct({
            who: LookupSource,
            newFree: sts.bigint(),
            newReserved: sts.bigint(),
        }),
        transfer: sts.enumStruct({
            dest: LookupSource,
            value: sts.bigint(),
        }),
        transfer_keep_alive: sts.enumStruct({
            dest: LookupSource,
            value: sts.bigint(),
        }),
    }
})

export const BabeCall: sts.Type<BabeCall> = sts.closedEnum(() => {
    return  {
        report_equivocation: sts.enumStruct({
            equivocationProof: BabeEquivocationProof,
            keyOwnerProof: KeyOwnerProof,
        }),
        report_equivocation_unsigned: sts.enumStruct({
            equivocationProof: BabeEquivocationProof,
            keyOwnerProof: KeyOwnerProof,
        }),
    }
})
// ...
export const AuthorshipCall: sts.Type<AuthorshipCall> = sts.closedEnum(() => {
    return  {
        set_uncles: sts.enumStruct({
            newUncles: sts.array(() => Header),
        }),
    }
})

export const AssetsCall: sts.Type<AssetsCall> = sts.closedEnum(() => {
    return  {
        burn: sts.enumStruct({
            assetId: AssetId,
            amount: TAssetBalance,
        }),
        mint: sts.enumStruct({
            assetId: AssetId,
            to: AccountId,
            amount: TAssetBalance,
        }),
        register: sts.enumStruct({
            symbol: AssetSymbol,
            name: AssetName,
            initialSupply: TAssetBalance,
            isMintable: sts.boolean(),
            isNft: sts.boolean(),
            optContentSrc: sts.option(() => ContentSource),
            optDesc: sts.option(() => Description),
        }),
        set_non_mintable: sts.enumStruct({
            assetId: AssetId,
        }),
        transfer: sts.enumStruct({
            assetId: AssetId,
            to: AccountId,
            amount: TAssetBalance,
        }),
    }
})
// ...
export const Type_43: sts.Type<Type_43> = sts.closedEnum(() => {
    return  {
        Assets: AssetsCall,
        Authorship: AuthorshipCall,
        Babe: BabeCall,
        Balances: BalancesCall,
        BridgeMultisig: BridgeMultisigCall,
        CeresLiquidityLocker: CeresLiquidityLockerCall,
        CeresStaking: CeresStakingCall,
        Council: CouncilCall,
        Currencies: CurrenciesCall,
        DEXAPI: DEXAPICall,
        Democracy: DemocracyCall,
        ElectionsPhragmen: ElectionsPhragmenCall,
        EthBridge: EthBridgeCall,
        Farming: FarmingCall,
        Faucet: FaucetCall,
        Grandpa: GrandpaCall,
        Identity: IdentityCall,
        ImOnline: ImOnlineCall,
        IrohaMigration: IrohaMigrationCall,
        LiquidityProxy: LiquidityProxyCall,
        MulticollateralBondingCurvePool: MulticollateralBondingCurvePoolCall,
        Multisig: MultisigCall,
        Offences: OffencesCall,
        Permissions: PermissionsCall,
        PoolXYK: PoolXYKCall,
        PswapDistribution: PswapDistributionCall,
        RandomnessCollectiveFlip: RandomnessCollectiveFlipCall,
        Referrals: ReferralsCall,
        Rewards: RewardsCall,
        Scheduler: SchedulerCall,
        Session: SessionCall,
        Staking: StakingCall,
        Sudo: SudoCall,
        System: SystemCall,
        Technical: TechnicalCall,
        TechnicalCommittee: TechnicalCommitteeCall,
        TechnicalMembership: TechnicalMembershipCall,
        Timestamp: TimestampCall,
        TradingPair: TradingPairCall,
        Utility: UtilityCall,
        VestedRewards: VestedRewardsCall,
        XSTPool: XSTPoolCall,
        XorFee: XorFeeCall,
    }
})